Filter and Sort in Notes, Images, Files
Add sort and filter in Notes, images and files and correct some related bugs in previous sort and filter.
Changed the Non Blocking counter in the StatusControlContext to use Interlocked after having a case come up with the straight property where the threading resulted in the wrong value being evaluated.