Content Editor Layout Change. About Support.
The action buttons in the content editor were originally in a tab - this was unfriendly esp. for cycles of updating and going to the site - moved to bottom and always visible.
Added support for an about page based on a Post Title - since in my imaginings there should be example one 'About' page went with 'magic' style where if you name a post 'About' you then have an about page...