Note Generation Fix. Clean Up.
Notes hadn't been added to the main window 'all' html generation - added.
Made a groupbox style so the groubox and border colors and thicknesses can match - didn't get the style everywhere (applying via style tag, not applying it as a default) but worked on getting it more places and on getting the resource dictionary into 'all' controls.
Deleted the older Windows for the lists - still a nice idea not to have the content lists stuck in tabs but not using at this point so eliminating.