Add Global Tool Install and Update Scripts. Clean ups from Script.
Added install and update scripts for XamlStyler and Jetbrains as Global Tools. Installing as global without warning would be inappropriate but I think the naming makes it clear what these are - in additional these tools are only for the additional scripts (at least for now) and not included in a build. Added cleanup script to run Jetbrains cleanup for **.cs files and XamlStyler.