Streamline Nuget Package References
From early development and habit from pre-Core .net the Nuget Package References were often needlessly included in project when an existing project reference would already provide access to the package. I have found this change worthwhile in other projects  - 99% of the time it won't matter but in the 1% of the time where packages have errors or warnings it can be helpful to not have the redundant package inclusions and multiple error/warning messages.