Fix Focus Bug with the Full Screen Message in the Status Control
While working on a site I noticed that when the Status Control Message screen is shown focus was apparently not moved to the Status Control and stayed with the UserControl launching the Message. This meant that if you accidentally pressed a button that triggered a (no longer visible) control to run commands (not intended or designed for) that the Status Control would end up in a state locked on the Status Log screen. This update is at least part 1 of addressing that by adding some focus triggers to the Status Control. This included changing the Message Button List to a structure that could support a 'default' button (eventually needed anyway!) that ties into what button has focus.

Also a small fix up in the Logging in StatusControlContext where some Serilog parameters had useless names.