Improve the Navigation Bar in Previews. Local Preview Window Title Fix.
The preview windows are for on site browsing only, the navigation bar now better reflects that with only the path showing in the open to edit section with the site url to the left. This also nicely hides the actual local host url.

Added better information to the main title bar in the stand alone preview window.