Improve Name and Position of Auto-Fix Filename Button - add to Photos and Images
The position of the Auto-Fix file button in the File Content Editor is improved moving it closer to the file name esp. since with it's coloration it also serves as a nice flag that the file name has an issue.

Added the Auto-Fix File Name button to the Image and Photo Content Editors.