Bug Fix in the Window Position Helper with screen ratio double applied - Initial Photo List Size
The Screen returns at least some values with the screen ratio already applied - in previous versions I had noticed occasional odd behaviour and realized that I was applying the screen ratio (the user scaling setting - set to 150% on my laptop) to the values that already applied that scaling. Fixed.
Adjusted the initial size of the PhotoListWindow to better show the sort buttons.