Photo Page Previous and Next Photo Links - Db Memory Access Bug Fix
Previous and Next photo links -not currently integrated into the changed html generation.
I had been seeing an occasional memory access exception - not fully diagnosed but caught the error this time and at a glance seemed like it could be the use of a List inside a parallel foreachasync - changed to a concurrent data structure.